system/exec: move exec utilities to their own package
The new package is a near drop-in replacement for the standard os/exec
package. NewCommand has been renamed to Command in order to match the
standard package. dnsmasq has switched to this new package, fixing error
reporting in its Destroy method.
